+++ example/Sorter.java
@@ -2,7 +2,7 @@ package example;
 
 public interface Sorter {
 
-    public void load(int[] elements);
+    public void load(double[] elements);
     public void sort();
     public String getPlan();
